requirements.txt
cached-property==1.5.2
cycler==0.10.0
DateTime==4.3
Django==2.2
django-crispy-forms==1.12.0
django-datetimepicker==3.14
fabio==0.11.0
h5py==3.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
mysqlclient==2.0.3
numpy==1.19.5
Pillow==8.2.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
silx==0.15.1
six==1.16.0
sqlparse==0.4.1
zope.interface==5.4.0
